Make tests run in temporary folder
