+++ src/daemon/ca/mod.rs
@@ -17,6 +17,7 @@ pub mod upgrades;
 
 pub use self::manager::CaManager;
 pub use self::manager::testbed_ca_handle;
+pub use self::status::CaStatusStore;
 pub use self::status::CaStatus;
