+++ src/daemon/ca/status.rs
@@ -580,53 +580,3 @@ impl CaStatus {
     }
 }
 
-
-//============ Tests =========================================================
-
-#[cfg(test)]
-mod tests {
-    use crate::commons::test;
-    use crate::constants::STATUS_NS;
-    use super::*;
-
-    #[test]
-    fn read_save_status() {
-        let source_dir_path_str =
-            "test-resources/status_store/migration-0.9.5/";
-        let source_dir_url =
-            Url::parse(&format!("local://{}", source_dir_path_str)).unwrap();
-
-        let source_store =
-            KeyValueStore::create(&source_dir_url, STATUS_NS).unwrap();
-
-        let test_storage_uri = test::mem_storage();
-        let status_kv_store =
-            KeyValueStore::create(&test_storage_uri, STATUS_NS).unwrap();
-
-        // copy the source KV store (files) into the test KV store (in memory)
-        status_kv_store.import(&source_store).unwrap();
-
-        // get the status for testbed before initialising a StatusStore
-        // using the copied the data - that will be done next and start
-        // a migration.
-        let testbed_status_key = Key::new_scoped(
-            Scope::from_segment(const { Segment::make("testbed") }),
-            Segment::parse("status.json").unwrap(),
-        );
-        let status_testbed_before_migration: CaStatus =
-            status_kv_store.get(&testbed_status_key).unwrap().unwrap();
-
-        // Initialise the StatusStore using the new (in memory) storage,
-        // and migrate the data.
-        let store =
-            CaStatusStore::create(&test_storage_uri, STATUS_NS).unwrap();
-        let testbed = CaHandle::from_str("testbed").unwrap();
-
-        // Get the migrated status for testbed and verify that it's equivalent
-        // to the status before migration.
-        let status_testbed_migrated = store.get_ca_status(&testbed);
-
-        assert_eq!(status_testbed_before_migration, status_testbed_migrated);
-    }
-}
-
